(refactoring) lift plugin related types from myst-common into myst-cli
