+++ packages/myst-cli/src/process/mdast.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { tic } from 'myst-cli-utils';
-import type { GenericParent, IExpressionResult, PluginUtils, References } from 'myst-common';
+import type { GenericParent, IExpressionResult, References } from 'myst-common';
 import { fileError, fileWarn, RuleId, slugToUrl } from 'myst-common';
 import type { PageFrontmatter } from 'myst-frontmatter';
 import { SourceFileKind } from 'myst-spec-ext';
@@ -35,10 +35,10 @@ import { unified } from 'unified';
 import { select, selectAll } from 'unist-util-select';
 import { VFile } from 'vfile';
 import { processPageFrontmatter, updateFileInfoFromFrontmatter } from '../frontmatter.js';
+import { PluginUtils } from '../plugins/index.js';
 import { selectors } from '../store/index.js';
-import type { ISession } from '../session/types.js';
-import { castSession } from '../session/cache.js';
-import type { RendererData } from '../transforms/types.js';
+import { type ISession, castSession } from '../session/index.js';
+import { type RendererData, rawDirectiveTransform } from '../transforms/index.js';
 
 import {
   checkLinksTransform,
@@ -68,20 +68,17 @@ import {
   transformLiftCodeBlocksInJupytext,
   transformMystXRefs,
 } from '../transforms/index.js';
-import type { ImageExtensions } from '../utils/resolveExtension.js';
-import { logMessagesFromVFile } from '../utils/logging.js';
+import { type ImageExtensions, logMessagesFromVFile, addEditUrl } from '../utils/index.js';
 import { combineCitationRenderers } from './citations.js';
 import { bibFilesInDir, selectFile } from './file.js';
 import { parseMyst } from './myst.js';
 import { kernelExecutionTransform, LocalDiskCache } from 'myst-execute';
 import type { IOutput } from '@jupyterlab/nbformat';
-import { rawDirectiveTransform } from '../transforms/raw.js';
-import { addEditUrl } from '../utils/addEditUrl.js';
 import {
   indexFrontmatterFromProject,
   manifestPagesFromProject,
   manifestTitleFromProject,
-} from '../build/utils/projectManifest.js';
+} from '../build/index.js';
 
 const LINKS_SELECTOR = 'link,card,linkBlock';
