+++ packages/myst-cli/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './addEditUrl.js';
 export * from './addWarningForFile.js';
 export * from './check.js';
 export * from './createTempFolder.js';
